chore: 공통으로 사용할 수 있는 것들 전부 공용 상수와 함수로 분리
- VALIDATION 관련 상수들
- id 검증용 함수들
